kubectl_explain_failure/requirements.txt
pytest
pyyaml
tox
